refactor(LineService) : Service 통합 및 재배치
